refactor(diagnose): split investigationEvidence into a directory
One file per concern, one adapter per tool family. Adding a tool now means a
new file under adapters/ and a line in the ADAPTERS registry, instead of
another few hundred lines in a file every other adapter is also editing.

  types.ts        the wire and projection types
  identity.ts     evidence refs, source ids, DOM ids
  parse.ts        payload coercion shared across adapters
  observations.ts ProjectionBuilder, relevance, limitations, observations
  adapters/       changes diagnose events helm issues logs
                  permissions prometheus resource topology
  builder.ts      the adapter registry and projectInvestigationEvidence
  index.ts        the public surface

index.ts re-exports exactly the 46 names the single file exported, and every
call site already imports "./investigationEvidence", which now resolves to the
directory — so no importer changes and no compatibility shim is left behind.
Symbols a sibling needs are exported at file level but deliberately kept out of
index.ts; the module's public surface is unchanged.

Pure move: no behaviour change. The token stream of the moved code is identical
to the original (39252 tokens, 1892 distinct, zero difference), the only edits
being added import/export lines and Prettier re-wrapping eleven signatures that
the export keyword pushed past the print width.
